requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
pandas==2.0.3
openpyxl==3.1.2
lxml==4.9.3
requests==2.31.0
aiofiles==23.1.0
python-docx==0.8.11 